refactor(Express adapter): refactored the adapter to be more configurable
